+++ nonebot/plugin.py
@@ -1,3 +1,4 @@
+import asyncio
 import os
 import re
 import sys
@@ -7,22 +8,35 @@ import importlib
 import contextlib
 from datetime import timedelta
 from types import ModuleType
-from typing import TYPE_CHECKING, Any, List, Set, Dict, Tuple, TypeVar, Union, Optional, Iterable, Callable, Type, overload
+from typing import TYPE_CHECKING, Any, Awaitable, Generator, List, Set, Dict, Tuple, TypeVar, Union, Optional, Iterable, Callable, Type, overload
 
 from .log import logger
 from nonebot import permission as perm
 from .command import Command, CommandManager, CommandSession
 from .notice_request import EventHandler, EventManager
 from .natural_language import NLProcessor, NLPManager
-from .typing import CommandName_T, CommandHandler_T, NLPHandler_T, NoticeHandler_T, Patterns_T, PermissionPolicy_T, RequestHandler_T
+from .helpers import separate_async_funcs
+from .typing import CommandName_T, CommandHandler_T, NLPHandler_T, NoticeHandler_T, Patterns_T, PermissionPolicy_T, PluginLifetimeHook_T, RequestHandler_T
 
 if TYPE_CHECKING:
     from .message import MessagePreprocessor
 
 
+class LifetimeHook:
+    """INTERNAL_API"""
+    __slots__ = ('func', 'timing')
+
+    def __init__(self, func: PluginLifetimeHook_T, timing: str):
+        if timing not in ('loading', 'unloaded'):
+            raise ValueError(f'Invalid timing "{timing}"')
+        self.func = func
+        self.timing = timing
+
+
 class Plugin:
     __slots__ = ('module', 'name', 'usage', 'userdata', 'commands', 'nl_processors',
-                 'event_handlers', 'msg_preprocessors')
+                 'event_handlers', 'msg_preprocessors', 'lifetime_hooks',
+                 '_load_future', '_command_args')
 
     def __init__(self,
                  module: ModuleType,
@@ -32,7 +46,8 @@ class Plugin:
                  commands: Set[Command] = ...,
                  nl_processors: Set[NLProcessor] = ...,
                  event_handlers: Set[EventHandler] = ...,
-                 msg_preprocessors: Set['MessagePreprocessor'] = ...):
+                 msg_preprocessors: Set['MessagePreprocessor'] = ...,
+                 lifetime_hooks: List[LifetimeHook] = ...):
         """Creates a plugin with no name, no usage, and no handlers."""
 
         self.module = module
@@ -47,6 +62,40 @@ class Plugin:
             event_handlers if event_handlers is not ... else set()
         self.msg_preprocessors: Set['MessagePreprocessor'] = \
             msg_preprocessors if msg_preprocessors is not ... else set()
+        self.lifetime_hooks: List[LifetimeHook] = \
+            lifetime_hooks if lifetime_hooks is not ... else []
+
+        self._load_future: Optional[asyncio.Future] = None
+        # backward compat without touching self.commands
+        self._command_args: Optional[Dict[
+            Command, Tuple[Union[Iterable[str], str], Patterns_T]]] = None
+
+    def __await__(self) -> Generator[None, None, Union['Plugin', None]]:
+        """Waits for the async (un)loading of the plugin."""
+        if self._load_future is not None:
+            try:
+                result = yield from self._load_future.__await__()
+                # if we are awaiting reload, self is stale plugin
+                # a reload call will return a new Plugin if successful
+                if result is not None:
+                    return (yield from result.__await__())
+                return self
+            except Exception:
+                return None
+            finally:
+                self._load_future = None
+        return self
+
+    def __del__(self):
+        # surpress unretrieved future exception warning
+        if self._load_future is not None:
+            self._load_future.cancel()
+
+    def _new_load_future(self) -> asyncio.Future:
+        if self._load_future is not None and not self._load_future.done():
+            self._load_future.set_exception(asyncio.CancelledError())
+        self._load_future = asyncio.get_event_loop().create_future()
+        return self._load_future
 
     class GlobalTemp:
         """INTERNAL API"""
@@ -56,6 +105,7 @@ class Plugin:
         nl_processors: Set[NLProcessor] = set()
         event_handlers: Set[EventHandler] = set()
         msg_preprocessors: Set['MessagePreprocessor'] = set()
+        lifetime_hooks: List[LifetimeHook] = []
         now_within_plugin: bool = False
 
         @classmethod
@@ -74,17 +124,22 @@ class Plugin:
             cls.nl_processors.clear()
             cls.event_handlers.clear()
             cls.msg_preprocessors.clear()
+            cls.lifetime_hooks.clear()
 
         @classmethod
         def make_plugin(cls, module: ModuleType):
-            return Plugin(module=module,
-                          name=getattr(module, '__plugin_name__', None),
-                          usage=getattr(module, '__plugin_usage__', None),
-                          userdata=getattr(module, '__plugin_userdata__', None),
-                          commands={cmd[0] for cmd in cls.commands},
-                          nl_processors={*cls.nl_processors},
-                          event_handlers={*cls.event_handlers},
-                          msg_preprocessors={*cls.msg_preprocessors})
+            p = Plugin(module=module,
+                       name=getattr(module, '__plugin_name__', None),
+                       usage=getattr(module, '__plugin_usage__', None),
+                       userdata=getattr(module, '__plugin_userdata__', None),
+                       commands={cmd[0] for cmd in cls.commands},
+                       nl_processors={*cls.nl_processors},
+                       event_handlers={*cls.event_handlers},
+                       msg_preprocessors={*cls.msg_preprocessors},
+                       lifetime_hooks=[*cls.lifetime_hooks])
+            # backward compat
+            p._command_args = {cmd[0]: (cmd[1], cmd[2]) for cmd in cls.commands}
+            return p
 
 
 class PluginManager:
@@ -300,19 +355,95 @@ class PluginManager:
             self.nlp_manager.switch_nlprocessor(processor, state)
 
 
-def _add_handlers_to_managers():
-    for cmd, aliases, patterns in Plugin.GlobalTemp.commands:
+def _add_handlers_to_managers(plugin: Plugin) -> None:
+    for cmd in plugin.commands:
         CommandManager.add_command(cmd.name, cmd)
-        # TODO: skip when command exists
-        CommandManager.add_aliases(aliases, cmd)
-        CommandManager.add_patterns(patterns, cmd)
-    for processor in Plugin.GlobalTemp.nl_processors:
+        if plugin._command_args is not None:
+            args = plugin._command_args[cmd]
+            CommandManager.add_aliases(args[0], cmd)
+            CommandManager.add_patterns(args[1], cmd)
+            plugin._command_args = None
+    for processor in plugin.nl_processors:
         NLPManager.add_nl_processor(processor)
-    for handler in Plugin.GlobalTemp.event_handlers:
+    for handler in plugin.event_handlers:
         EventManager.add_event_handler(handler)
     from .message import MessagePreprocessorManager  # avoid import cycles
-    for mp in Plugin.GlobalTemp.msg_preprocessors:
+    for mp in plugin.msg_preprocessors:
         MessagePreprocessorManager.add_message_preprocessor(mp)
+
+
+def _run_async_func_by_environ(func: Callable[..., Awaitable[Any]]) -> None:
+    """
+    run an async func depending on whether we are currently in a running
+    event loop (inside a another async function)
+    """
+    try:
+        loop = asyncio.get_running_loop()
+    except RuntimeError:  # There is no current event loop..
+        loop = None
+    if loop and loop.is_running():
+        loop.create_task(func())
+    else:
+        # not using asyncio.run() because it can be called only once (ideally)
+        asyncio.get_event_loop().run_until_complete(func())
+
+
+def _clean_up_module(module_path: str):
+    for module in [m for m in sys.modules.keys() if m.startswith(module_path)]:
+        del sys.modules[module]
+
+
+def _load_plugin(module_path: str, act: str) -> Optional[Plugin]:
+    if PluginManager.get_plugin(module_path) is not None:
+        warnings.warn(f"Plugin {module_path} already exists")
+        return
+
+    imported = False
+    try:
+        with Plugin.GlobalTemp.enter_plugin():
+            module = importlib.import_module(module_path)
+        imported = True
+        plugin = Plugin.GlobalTemp.make_plugin(module)
+
+        sync_loaders, async_loaders = separate_async_funcs(
+            f.func for f in plugin.lifetime_hooks if f.timing == 'loading'
+        )
+        for f in sync_loaders:
+            f()
+        if not async_loaders:
+            # at this point, GlobalTemp and plugin object ^ have same contents
+            _add_handlers_to_managers(plugin)
+            PluginManager.add_plugin(module_path, plugin)
+            logger.info(f'Succeeded to {act} "{module_path}"')
+            return plugin
+        # continue the async loading after this functions returns
+        fut = plugin._new_load_future()
+
+        async def new_loader():
+            try:
+                for f in async_loaders:
+                    await f()
+                # but not necessarily here
+                _add_handlers_to_managers(plugin)
+                PluginManager.add_plugin(module_path, plugin)
+                logger.info(f'Succeeded to {act} "{module_path}"')
+                fut.set_result(None)
+            except Exception as e:
+                if imported:
+                    _clean_up_module(module_path)
+                fut.set_exception(e)
+                logger.error(f'Failed to run loading hooks when {act}ing '
+                             f'"{module_path}" asynchronously, error: {e}')
+                logger.exception(e)
+
+        _run_async_func_by_environ(new_loader)
+        return plugin
+    except Exception as e:
+        if imported:
+            _clean_up_module(module_path)
+        logger.error(f'Failed to {act} "{module_path}", error: {e}')
+        logger.exception(e)
+        return None
 
 
 def load_plugin(module_path: str) -> Optional[Plugin]:
@@ -322,26 +453,71 @@ def load_plugin(module_path: str) -> Optional[Plugin]:
         module_path (str): path of module to import
     
     Returns:
-        Optional[Plugin]: Plugin object loaded
+        Optional[Plugin]: Plugin object loaded, which can be awaited if
+                          the caller wishes to wait for async loading
+                          callbacks if there is any, or None loading fails
     """
-    try:
-        with Plugin.GlobalTemp.enter_plugin():
-            module = importlib.import_module(module_path)
-        plugin = Plugin.GlobalTemp.make_plugin(module)
+    return _load_plugin(module_path, 'import and load')
 
-        # at this point, GlobalTemp and plugin object ^ have same contents
-        _add_handlers_to_managers()
 
-        PluginManager.add_plugin(module_path, plugin)
-        logger.info(f'Succeeded to import "{module_path}"')
-        return plugin
-    except Exception as e:
-        logger.error(f'Failed to import "{module_path}", error: {e}')
-        logger.exception(e)
+def _unload_plugin(module_path: str,
+                   kont: Optional[Callable[[], Any]]) -> Optional[Plugin]:
+    plugin = PluginManager.get_plugin(module_path)
+    if not PluginManager.remove_plugin(module_path) or plugin is None:
+        # second condition is useless. just pass type check
         return None
 
+    sync_unloaders, async_unloaders = separate_async_funcs(
+        f.func for f in plugin.lifetime_hooks if f.timing == 'unloaded'
+    )
 
-def unload_plugin(module_path: str) -> bool:
+    # docs say behavior is undefined if unloaders raise, but the case is still
+    # handled like this under the hood
+    error = False
+
+    try:
+        for f in sync_unloaders:
+            f()
+    except Exception as e:
+        error = True
+        logger.error(f'Failed to run unloading hooks when unloading '
+                     f'"{module_path}", error: {e}. Remaining hooks are not continued.')
+        logger.exception(e)
+
+    def after_cbs():
+        _clean_up_module(module_path)
+        if error:
+            logger.info(f'Unloaded "{module_path}" with error')
+        else:
+            logger.info(f'Succeeded to unload "{module_path}"')
+
+    if not async_unloaders:
+        after_cbs()
+        if kont is not None:
+            return kont()
+        return plugin  # this is not None
+    # continue the async unloading after this functions returns
+    fut = plugin._new_load_future()
+
+    async def new_unloader():
+        try:
+            for f in async_unloaders:
+                await f()
+        except Exception as e:
+            nonlocal error
+            error = True
+            logger.error(f'Failed to run unloading hooks when unloading '
+                         f'"{module_path}" asynchronously, error: {e}.'
+                         'Remaining hooks are not continued.')
+            logger.exception(e)
+        after_cbs()
+        fut.set_result(kont() if kont is not None else None)
+
+    _run_async_func_by_environ(new_unloader)
+    return plugin
+
+
+def unload_plugin(module_path: str) -> Optional[Plugin]:
     """Unloads a plugin.
 
     This deletes its entry in sys.modules if present. However, if the module
@@ -352,35 +528,25 @@ def unload_plugin(module_path: str) -> bool:
         module_path (str): import path to module, which is already imported
 
     Returns:
-        bool: Plugin successfully unloaded
+        Optional[Plugin]: Stale Plugin (which can be awaited if the caller
+                          wishes to wait for async unloaded callbacks if there
+                          is any) if it was unloaded, None if it were not
+                          loaded
     """
-    result = PluginManager.remove_plugin(module_path)
-    if result:
-        for module in [m for m in sys.modules.keys() if m.startswith(module_path)]:
-            del sys.modules[module]
-        logger.info(f'Succeeded to unload "{module_path}"')
-    return result
+    return _unload_plugin(module_path, None)
 
 
 def reload_plugin(module_path: str) -> Optional[Plugin]:
-    if not unload_plugin(module_path):  # the one more mysterious log info seems valid..
-        return None
+    """A combination of unload and load of a plugin.
+    
+    Args:
+        module_path (str): import path to module, which is already imported
 
-    try:
-        with Plugin.GlobalTemp.enter_plugin():
-            # NOTE: consider importlib.reload()
-            module = importlib.import_module(module_path)
-        plugin = Plugin.GlobalTemp.make_plugin(module)
-
-        _add_handlers_to_managers()
-
-        PluginManager.add_plugin(module_path, plugin)
-        logger.info(f'Succeeded to reload "{module_path}"')
-        return plugin
-    except Exception as e:
-        logger.error(f'Failed to reload "{module_path}", error: {e}')
-        logger.exception(e)
-        return None
+    Returns:
+        Optional[Plugin]: The return value is special, please see the doc
+    """
+    # NOTE: consider importlib.reload()
+    return _unload_plugin(module_path, lambda: _load_plugin(module_path, 'reload'))
 
 
 def load_plugins(plugin_dir: str, module_prefix: str) -> Set[Plugin]:
@@ -431,6 +597,24 @@ def get_loaded_plugins() -> Set[Plugin]:
     :return: a set of Plugin objects
     """
     return set(PluginManager._plugins.values())
+
+
+def on_plugin(timing: str) -> Callable[[PluginLifetimeHook_T], PluginLifetimeHook_T]:
+    """
+    Decorator to register a function as a callback for plugin lifetime events.
+
+    Args:
+        timing (str): Either 'loading' or 'unloaded'
+    """
+    def deco(func: PluginLifetimeHook_T):
+        if Plugin.GlobalTemp.now_within_plugin:
+            hk = LifetimeHook(func, timing)
+            Plugin.GlobalTemp.lifetime_hooks.append(hk)
+        else:
+            raise RuntimeError('Cannot register a lifetime hook outside a plugin')
+        return func
+
+    return deco
 
 
 def on_command(
@@ -642,6 +826,7 @@ __all__ = [
     'load_plugins',
     'load_builtin_plugins',
     'get_loaded_plugins',
+    'on_plugin',
     'on_command',
     'on_natural_language',
     'on_notice',